requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.1.1
bcrypt==5.0.0
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0